requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.11.0
python-dotenv==1.0.1
pydantic==2.9.2
streamlit==1.39.0
requests==2.32.3
pyperclip==1.9.0
python-multipart==0.0.12
